requirements.txt
anyio==4.0.0
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-lru==2.0.4
attrs==23.1.0
azure-core==1.29.5
azure-common==1.1.28
azure-communication-email==1.0.0
azure-mgmt-core==1.4.0
azure-storage-blob==12.18.3
Babel==2.13.1
beautifulsoup4==4.12.2
black==23.11.0
bleach==6.1.0
boto3==1.28.44
botocore==1.31.44
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
comm==0.1.4
contextlib2==21.6.0
coverage==7.3.2
cryptography==41.0.4
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
django-cors-headers==4.3.0
django-ninja==0.22.2
django-ninja-extra==0.19.8
django-ninja-jwt==5.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-storages
executing==2.0.1
Faker==20.1.0
fastjsonschema==2.18.1
fqdn==1.5.1
gunicorn==21.2.0
idna==3.4
iniconfig==2.0.0
injector==0.21.0
ipykernel==6.26.0
ipython==8.17.2
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-events==0.8.0
jupyter-lsp==2.2.0
jupyter_client==8.5.0
jupyter_core==5.5.0
jupyter_server==2.9.1
jupyter_server_terminals==0.4.4
jupyterlab==4.0.8
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mistune==3.0.2
msrest==0.7.1
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.10.0
nbformat==5.9.2
nest-asyncio==1.5.8
ninja==1.11.1
notebook_shim==0.2.3
oauthlib==3.2.2
overrides==7.4.0
packaging==23.1
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
Pillow==10.0.0
platformdirs==3.11.0
pluggy==1.3.0
prometheus-client==0.18.0
prompt-toolkit==3.0.39
psutil==5.9.6
psycopg2==2.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.12
Pygments==2.16.1
PyJWT==2.8.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-pdb==0.3.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.12.0
s3transfer==0.6.2
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.3.0
urllib3==1.26.16
wcwidth==0.2.9
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
whitenoise==6.6.0
holidays==0.39
numpy==1.26.4
pvlib==0.10.3
PuLP==2.8.0
python-dotenv==1.0.0
python-slugify==8.0.4